refactor creator_ui into modular files
